qa-crawler/backend/requirements.txt
playwright
beautifulsoup4
python-dotenv
asyncio
flask
openai
requests
flask-cors
dirtyjson
imagehash